Update captions and descriptions for Tamils section
I changed the wording from "Indian Tamils" to "Malaiyaha Tamils" to represent their identity in a non-discriminatory way. As a person who lives outside of Sri Lanka, I wasn't aware of the politically correct way of referencing the Tamil population in the Sri Lankan highlands. My brother, a human rights activist, pointed out that my wording was insensitive and wrong. I apologise for this.